+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add right click context menu items to bindings in the list view which can be used to copy, paste, duplciate, and remove them
 - Add a new item template creation window which lists commonly used binding configurations
 - Add the ability to create groups/folders to organize bindings
+- Add a class load option
 - Remember the selected binding page (action, targets, conditions) when switching between bindings
 
 ### Changed
@@ -25,6 +26,7 @@
 - Fix default dropdown value for the War Mode condition
 - Fix ElvUI skinning if the "Ace3" option is disabled within ElvUI
 - Fix binding list not always remembering the selected binding after deleting one
+- Fix PvP talents being listed multiple times
 
 ## [0.9.3] - 2020-10-22